abstracts Marker, getBounds
