+++ src/components/index.js
@@ -3,3 +3,4 @@
 export * from './Navbar'
 export * from './Footer'
 export * from './Logo'
+export * from './Marker'
