+++ src/utils/index.js
@@ -1,4 +1,3 @@
 // utility function import/export file
-import { renderRating } from './renderRating'
-
-export { renderRating }
+export * from './renderRating'
+export * from './getBounds'
